third_party/blink/web_tests/virtual/select-parser-relaxation/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-input-keyboard-behavior.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Keyboard behavior of <input> in <select>.
  assert_equals: appearance: base-select must be supported for this test to run. expected "base-select" but got "auto"
Harness: the test ran to completion.
